Add Microsoft Defender collector bot
This bot is quite a mess - it horribly abuses the "extra." namespace
in messages, it hard-codes processing decisions, the error handling
is not great, and it cannot do OAuth2 token refresh, so it needs to
re-authenticate for each call.

Still, it does what it's supposed to do, and maybe someone else can
use it, even if it is never accepted in IntelMQ upstream.