Add dynamic mocking of API endpoint
Since every API call goes through the same function,
requests_oauthlib.OAuth2Session.get(), we need logic in the mock to
select the corret response if we're going to test anything more
complicated than a static error.

Create a class Mock_API_Endpoint that selects its response based
on the prefix of the URL used to call it.