Add directories for split Defender functionality
The Defender collector bot is a bit too far removed from IntelMQ's
design to really fit. It centralises too many features and is
too specific to our use case.

The bot will be split in four:
1) collector, that fetches alerts.
2) parser, that decodes JSON.
3) file expert, that makes API queries for files in the alert.
4) advanced hunting expert, that can make arbitary API queries. We
will use it to ask for affected users.

As a first step in this redesign, create the directories that the
new bots will live in.